AI_bE/requirements.txt
fastapi
uvicorn
openai
gtts
google-cloud-speech
websockets
python-dotenv
httpx
aiohttp
feedparser
wikipedia
websockets
beautifulsoup4
google-api-python-client
uvicorn
edge-tts>=6.1.10
